+++ MyIdDatabase/MyIdDb.sqlproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <Name>MyIdDatabase</Name>
+    <Name>MyIdDb</Name>
     <SchemaVersion>2.0</SchemaVersion>
     <ProjectVersion>4.1</ProjectVersion>
     <ProjectGuid>{526e316d-fdf0-447d-b281-6191164336d2}</ProjectGuid>
