+++ MyIdLibrary/DataAccess/DbConnection.cs
@@ -1,0 +1,5 @@
+﻿namespace MyIdLibrary.DataAccess;
+
+public class DbConnection
+{
+}
