+++ MyIdLibrary/Models/SecretModel.cs
@@ -1,0 +1,5 @@
+﻿namespace MyIdLibrary.Models;
+
+public class SecretModel
+{
+}
