+++ MyIdWeb/GlobalUsings.cs
@@ -1,0 +1,2 @@
+﻿global using MyIdLibrary.Models;
+global using MyIdLibrary.DataAccess;
