+++ MyIdWeb/MyIdWeb.csproj
@@ -14,4 +14,12 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="8.0.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="wwwroot\images\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MyIdLibrary\MyIdLibrary.csproj" />
+  </ItemGroup>
+
 </Project>
